pyproject.toml
[project]
name = "pantry-bot"
version = "0.1.0"
description = "Telegram bot for managing a shared home pantry, with LLM-powered parsing and recipe suggestions"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=22.7",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "python-dateutil>=2.9.0",
]

[project.scripts]
pantry-bot = "pantry_bot.main:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["src"]
